+++ api/c/src/routines/potrf.cc
@@ -3,7 +3,7 @@
 /*   potrf.cc                                                     .-*-.       */
 /*                                                              .'* *.'       */
 /*   Created: 2024/07/09 11:22:22 by Romain Pereira          __/_*_*(_        */
-/*   Updated: 2025/12/15 19:15:30 by Romain PEREIRA         / _______ \       */
+/*   Updated: 2026/01/08 16:32:34 by Romain PEREIRA         / _______ \       */
 /*                                                          \_)     (_/       */
 /*   License: CeCILL-C                                                        */
 /*                                                                            */
@@ -24,7 +24,7 @@ int
 xkblas_£potrf_tile_async(
     int uplo,
     int n,
-    TYPE * A, const size_t Atm, const size_t Atn, const size_t Amb, const size_t Anb, size_t lda,
+    TYPE * A, const int Atm, const int Atn, const int Amb, const int Anb, int lda,
     xkrt_device_global_id_t device_global_id
 ) {
     return xkblas_get()->potrf_tile_async<xkblas_precision_t::££>(uplo, n, A, Atm, Atn, Amb, Anb, lda, device_global_id);
